+++ app/helpers/energy_tariffs_helper.rb
@@ -141,4 +141,14 @@ module EnergyTariffsHelper
   def settings(charge_type)
     EnergyTariffCharge.charge_types[charge_type.to_sym] || {}
   end
+
+  def any_smart_meters?(school)
+    school.meters.dcc.any?
+  end
+
+  def smart_meter_tariffs(meter)
+    smart_meter_tariff_attributes = meter.smart_meter_tariff_attributes
+    return [] unless smart_meter_tariff_attributes
+    smart_meter_tariff_attributes[:accounting_tariff_generic]
+  end
 end
